Implemented generic page allocator
The entry point now uses it properly. I am preparing to write what is perhaps the most challenging algorithm for me yet, to parse the memory map in the smartest way I can think of. The page allocator class, which is not yet complete (but it is perfectly functional), easily allows allocation spaces to be given priority (each allocation spaces is essentially a node in a doubly-linked list) and will make memory hotplug effortless to implement.