Updated object allocator
Fixed a bug that would've left the linkage lock engaged when the first
pool cannot be released.
Also added more options for dealing with pool releasing behavior.

The release options enum is defined in `object_allocator_pools.hpp`.
It's not defined in `object_allocator_hbase.inc` because there's nothing
unique to this enum and would result in a duplicate when both the
single-consumer and multi-consumer versions are used.